Now inspects @Validate annotations and automatically includes those properties in the allowed list.
Changed the API so that one instance is created and cached for each ActionBean class that contains all the information relevant to that class.

Moved the ActionBeanContext security check into BindingPolicyManager.

Fixed a bug in the glob to regex translation for nested properties.

Resolved STS-398: @StrictBinding does not log the properties it blocks.

Also did some additional cleanup.