STS-435: Change clean URLs not to rely on internal forwards
When the resolver finds an ActionBean that already exists in the request scope, it must ensure that the bean's context has a reference to exactly the same request object that the dispatcher servlet is using at the time. This is important for processing of includes and forwards. Specifically, this is essential to make URI-embedded request parameters visible during an include or forward.