Remove actionBean request attribute when stack is empty.
Tomcat's crossContext configuration item means that a single request may be forwarded to multiple servlets. This clears the actionBean request attribute when the stack is empty so that the next servlet is not polluted with an action bean from the previous servlet.

Some of the tests assumed that the last action bean to be restored from the stack would remain in the actionBean request attribute. To resolve this, a new interceptor was introduced for testing which saves out the last action bean prior to it being cleared.