Prepend contextPath to the action bean's session attribute.
Tomcat's crossContext configuration item means that a single request may be forwarded to multiple servlets. If multiple instances of stripes are running in different webapps in the same container, with this configuration it is possible that they'll share a context, and thus the URL binding may collide.

The MockRoundtrip class expected the attribute key to be the URL Binding, so this was updated to align with this change.